Remove Client class from transmartApp (already moved to transmart-oauth)
